+++ programs/dynamic-bonding-curve/src/instructions/migration/ix_withdraw_migration_fee.rs
@@ -6,7 +6,7 @@ use crate::{
     const_pda,
     event::EvtWithdrawMigrationFee,
     state::{MigrationFeeDistribution, CREATOR_MIGRATION_FEE_MASK, PARTNER_MIGRATION_FEE_MASK},
-    token::transfer_token_from_pool_authority,
+    token::{transfer_token_from_pool_authority, validate_transfer_fee_is_zero},
     ConfigAccountLoader, PoolAccountLoader, PoolError,
 };
 
@@ -126,6 +126,8 @@ pub fn handle_withdraw_migration_fee(
         pool.update_withdraw_migration_fee(mask);
         creator_migration_fee
     };
+
+    validate_transfer_fee_is_zero(&ctx.accounts.quote_mint.to_account_info())?;
 
     transfer_token_from_pool_authority(
         ctx.accounts.pool_authority.to_account_info(),