+++ programs/dynamic-bonding-curve/src/state/mod.rs
@@ -11,3 +11,5 @@ pub mod virtual_pool_metadata;
 pub use virtual_pool_metadata::*;
 pub mod operator;
 pub use operator::*;
+pub mod token_badge;
+pub use token_badge::*;