+++ tests/config_pool_creation_fee.tests.ts
@@ -301,7 +301,7 @@ async function createConfigAccount(
     totalTokenSupply,
     initialMarketcap,
     migrationMarketcap,
-    0,
+    1, // damm v2
     tokenBaseDecimal,
     tokenQuoteDecimal,
     0,