+++ tests/instructions/userInstructions.ts
@@ -49,6 +49,8 @@ export type CreatePoolSplTokenParams = {
   quoteMint: PublicKey;
   config: PublicKey;
   instructionParams: InitializePoolParameters;
+  tokenQuoteProgram?: PublicKey;
+  tokenBadge?: PublicKey;
 };
 
 export type CreatePoolToken2022Params = CreatePoolSplTokenParams;
@@ -93,9 +95,14 @@ export async function createInitializePoolWithSplTokenIx(
       quoteVault,
       mintMetadata,
       metadataProgram: METAPLEX_PROGRAM_ID,
-      tokenQuoteProgram: TOKEN_PROGRAM_ID,
+      tokenQuoteProgram: params.tokenQuoteProgram ?? TOKEN_PROGRAM_ID,
       tokenProgram,
     })
+    .remainingAccounts(
+      params.tokenBadge
+        ? [{ pubkey: params.tokenBadge, isSigner: false, isWritable: false }]
+        : []
+    )
     .instruction();
 
   return {
@@ -154,9 +161,14 @@ export async function createPoolWithToken2022(
       poolAuthority,
       baseVault,
       quoteVault,
-      tokenQuoteProgram: TOKEN_PROGRAM_ID,
+      tokenQuoteProgram: params.tokenQuoteProgram ?? TOKEN_PROGRAM_ID,
       tokenProgram: TOKEN_2022_PROGRAM_ID,
     })
+    .remainingAccounts(
+      params.tokenBadge
+        ? [{ pubkey: params.tokenBadge, isSigner: false, isWritable: false }]
+        : []
+    )
     .transaction();
 
   transaction.add(
@@ -202,9 +214,14 @@ export async function createPoolWithToken2022TransferHook(
       baseVault,
       quoteVault,
       transferHookProgram,
-      tokenQuoteProgram: TOKEN_PROGRAM_ID,
+      tokenQuoteProgram: params.tokenQuoteProgram ?? TOKEN_PROGRAM_ID,
       tokenProgram: TOKEN_2022_PROGRAM_ID,
     })
+    .remainingAccounts(
+      params.tokenBadge
+        ? [{ pubkey: params.tokenBadge, isSigner: false, isWritable: false }]
+        : []
+    )
     .transaction();
 
   transaction.add(
@@ -416,9 +433,10 @@ export async function swap(
   const isInputBaseMint = inputTokenMint.equals(poolState.baseMint);
 
   const quoteMint = isInputBaseMint ? outputTokenMint : inputTokenMint;
+  const tokenQuoteProgram = svm.getAccount(quoteMint).owner;
   const [inputTokenProgram, outputTokenProgram] = isInputBaseMint
-    ? [tokenBaseProgram, TOKEN_PROGRAM_ID]
-    : [TOKEN_PROGRAM_ID, tokenBaseProgram];
+    ? [tokenBaseProgram, tokenQuoteProgram]
+    : [tokenQuoteProgram, tokenBaseProgram];
 
   const preInstructions: TransactionInstruction[] = [];
   const postInstructions: TransactionInstruction[] = [];
@@ -475,7 +493,7 @@ export async function swap(
       quoteMint,
       payer: payer.publicKey,
       tokenBaseProgram,
-      tokenQuoteProgram: TOKEN_PROGRAM_ID,
+      tokenQuoteProgram,
       referralTokenAccount,
     })
     .remainingAccounts(
@@ -549,9 +567,10 @@ export async function swapWithTransferHook(
 
   const isInputBaseMint = inputTokenMint.equals(poolState.baseMint);
   const quoteMint = isInputBaseMint ? outputTokenMint : inputTokenMint;
+  const tokenQuoteProgram = svm.getAccount(quoteMint).owner;
   const [inputTokenProgram, outputTokenProgram] = isInputBaseMint
-    ? [tokenBaseProgram, TOKEN_PROGRAM_ID]
-    : [TOKEN_PROGRAM_ID, tokenBaseProgram];
+    ? [tokenBaseProgram, tokenQuoteProgram]
+    : [tokenQuoteProgram, tokenBaseProgram];
 
   const preInstructions: TransactionInstruction[] = [];
   const postInstructions: TransactionInstruction[] = [];
@@ -622,7 +641,7 @@ export async function swapWithTransferHook(
       quoteMint,
       payer: payer.publicKey,
       tokenBaseProgram,
-      tokenQuoteProgram: TOKEN_PROGRAM_ID,
+      tokenQuoteProgram,
       referralTokenAccount,
     })
     .remainingAccounts([