+++ tests/migrate_to_damm_v2_with_vesting.tests.ts
@@ -199,7 +199,7 @@ async function fullFlow(
     totalTokenSupply,
     initialMarketcap,
     migrationMarketcap,
-    0,
+    1, // damm v2
     tokenBaseDecimal,
     tokenQuoteDecimal,
     0,